updated main.py
Added run configuration
Added dev requirements
