+++ src/airflow/dev-requirements.txt
@@ -1,4 +1,13 @@
 apache-airflow
 apache-airflow-providers-slack
+confluent-avro
 google-cloud-bigquery
 google-cloud-storage
+pyarrow
+sentry-sdk
+pandas
+requests
+simplejson
+python-dotenv
+kafka-python
+numpy
